Add files via upload
A desert background I started making for an assignment at school. We were asked to make a background of our choice, then add three roads each with their own stop sign. It was also required that at least one stop sign blinks. This was a group assignment and I was voted team leader. I asked that my three team members work on functions for a sun, clouds, and tumbleweeds for the scenery.